Remove find query method
This removes the find query method. The Montage API will not take a
'get' query alongside an empty 'filter'. To reduce complexity, I removed
the method. I also updated the document function in the client to work
on the new endpoint. I also added some sugar to the auth method that
sets the token on a successful authorization.